running better

+++ Polynomial.cpp
@@ -22,13 +22,13 @@ void Polynomial<ItemType>::insert( const ItemType& newEntry )    // inserts term
 template<class ItemType>
 int Polynomial<ItemType>::solve(int x)    // solves polynomial and returns the sum
 {
-    double sum = 0;
+    int sum = 0;
     Node<ItemType>* curPtr = headPtr;
 
     while (curPtr != nullptr)    //traverses through linked list
     {
-        int c = (curPtr->getItem()).first;    //gets the constant out of the term
-        unsigned k = (curPtr->getItem()).second;    //gets the exponent out of the term
+        double c = (curPtr->getItem()).first;    //gets the constant out of the term
+        int k = (curPtr->getItem()).second;    //gets the exponent out of the term
         
         sum = sum+(c*pow(x,k));    //sums
